requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
ExifRead==3.3.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.11.0.86
piexif==1.1.3
pillow==11.1.0
pymediainfo==7.0.1
qrcode==8.0
Werkzeug==3.1.3
